eval/requirements.txt
pandas
pyyaml
google-generativeai
jiwer
librosa
soundfile
onnxruntime
google-cloud-bigquery
google-cloud-storage
psycopg2-binary
python-dotenv
